Changes name to InputParser and cleans up unused code.
